no Warnings are exists

+++ src/Watchable.cpp
@@ -35,7 +35,7 @@ Episode::Episode(long id, const std::string &seriesName, int length, int season,
 Episode::Episode(Episode &other):Watchable(other.getId(),other.getLength(),other.getTags()),seriesName(other.getSeriesName()),season(other.getSeason()),episode(other.getEpisode()), nextEpisodeId(other.nextEpisodeId){}
 Watchable* Episode::getNextWatchable(Session &s) const {
     const std::vector<Watchable*>& content=s.getContent();
-    if(nextEpisodeId<content.size() && content.at(nextEpisodeId)->isEpisode())
+    if((size_t)nextEpisodeId<content.size() && content.at(nextEpisodeId)->isEpisode())
         return content.at(nextEpisodeId);
     else
         return nullptr;
